Refactor to use `short` instead of `int` for word indexes and matrices to optimize memory usage (#4)

+++ src/WordleStartOptimizer/Models/CandidateSet.cs
@@ -7,6 +7,6 @@ public readonly struct CandidateSet
 {
     private string DisplayString => string.Join(", ", WordIndexes.Select(x => Data.ValidGuesses[x]));
 
-    public required int[]  WordIndexes { get; init; }
+    public required short[]  WordIndexes { get; init; }
     public required double PreScore    { get; init; }
 }
